+++ CHANGELOG.rst
@@ -6,7 +6,15 @@ Current (0.2.x)
 
 - Upgrade Dependencies (#61)
 
+- CLI: export-labels
+
 - Make clustering more explorative by removing a fraction of the dataset that is closest to already validated clusters (#72)
+
+
+0.2.1
+=====
+
+(Multiple improvements without a pull request.)
 
 
 0.2.0